# openapi.json
{
  "components": {
    "schemas": {
...
      "AssetRefDTO": {
        "additionalProperties": false,
        "description": "Stable object reference; never carries a signed URL.",
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "provider_key": {
            "title": "Provider Key",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_key",
          "object_key",
          "mime_type",
          "byte_size",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "AssetRefDTO",
        "type": "object"
      },
...
      "AssignRoleBody": {
        "additionalProperties": false,
        "properties": {
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "default": "identity",
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "subject_id"
        ],
        "title": "AssignRoleBody",
        "type": "object"
      },
...
      "CapabilityDTO": {
        "additionalProperties": false,
        "properties": {
          "keys": {
            "items": {
              "type": "string"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "keys"
        ],
        "title": "CapabilityDTO",
        "type": "object"
      },
...
      "CreateRoleBody": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CreateRoleBody",
        "type": "object"
      },
...
      "CreateUploadIntentInput": {
        "additionalProperties": false,
        "properties": {
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256"
          },
          "content_length_max": {
            "exclusiveMinimum": 0.0,
            "maximum": 104857600.0,
            "title": "Content Length Max",
            "type": "integer"
          },
          "mime_types": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Mime Types",
            "type": "array"
          },
          "provider_key": {
            "title": "Provider Key",
            "type": "string"
          }
        },
        "required": [
          "provider_key",
          "mime_types",
          "content_length_max"
        ],
        "title": "CreateUploadIntentInput",
        "type": "object"
      },
      "CreateUploadIntentResult": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "intent_id": {
            "title": "Intent Id",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "intent_id",
          "object_key",
          "upload_url",
          "expires_at"
        ],
        "title": "CreateUploadIntentResult",
        "type": "object"
      },
...
      "FinalizeResultDTO": {
        "additionalProperties": false,
        "description": "Ack returned by FinalizeAsset; the asset becomes observable after\nthe finalize workflow executes.",
        "properties": {
          "intent_id": {
            "title": "Intent Id",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "state": {
            "default": "pending",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "intent_id",
          "object_key"
        ],
        "title": "FinalizeResultDTO",
        "type": "object"
      },
      "GrantSummary": {
        "additionalProperties": false,
        "description": "Role grants for one subject (used by read surfaces).",
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "subject_type",
          "subject_id"
        ],
        "title": "GrantSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterExternalAssetInput": {
        "additionalProperties": false,
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "bucket": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket"
          },
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mime_type": {
            "maxLength": 200,
            "title": "Mime Type",
            "type": "string"
          },
          "object_key": {
            "maxLength": 500,
            "title": "Object Key",
            "type": "string"
          },
          "provider_key": {
            "title": "Provider Key",
            "type": "string"
          }
        },
        "required": [
          "provider_key",
          "object_key",
          "mime_type",
          "byte_size"
        ],
        "title": "RegisterExternalAssetInput",
        "type": "object"
      },
...
      "ResolvedAssetUrlDTO": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "url",
          "expires_in_seconds"
        ],
        "title": "ResolvedAssetUrlDTO",
        "type": "object"
      },
      "RoleDTO": {
        "additionalProperties": false,
        "properties": {
          "capability_keys": {
            "items": {
              "type": "string"
            },
            "title": "Capability Keys",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "system": {
            "default": false,
            "title": "System",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "RoleDTO",
        "type": "object"
      },
...
      "UpdateAssetMetadataInput": {
        "additionalProperties": false,
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "title": "UpdateAssetMetadataInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "aiya-cms (cms)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/assets": {
      "post": {
        "operationId": "register_external_api_v1_admin_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterExternalAssetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRefDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register External",
        "tags": [
          "admin",
          "admin-assets"
        ]
      }
    },
    "/api/v1/admin/assets/upload-intents": {
      "post": {
        "operationId": "create_upload_intent_api_v1_admin_assets_upload_intents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadIntentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUploadIntentResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Upload Intent",
        "tags": [
          "admin",
          "admin-assets"
        ]
      }
    },
    "/api/v1/admin/assets/upload-intents/{intent_id}/finalize": {
      "post": {
        "operationId": "finalize_upload_api_v1_admin_assets_upload_intents__intent_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizeResultDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finalize Upload",
        "tags": [
          "admin",
          "admin-assets"
        ]
      }
    },
    "/api/v1/admin/assets/{asset_id}": {
      "delete": {
        "operationId": "delete_asset_api_v1_admin_assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Asset",
        "tags": [
          "admin",
          "admin-assets"
        ]
      },
      "get": {
        "operationId": "get_asset_api_v1_admin_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRefDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Asset",
        "tags": [
          "admin",
          "admin-assets"
        ]
      },
      "patch": {
        "operationId": "update_metadata_api_v1_admin_assets__asset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetMetadataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRefDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Metadata",
        "tags": [
          "admin",
          "admin-assets"
        ]
      }
    },
    "/api/v1/admin/assets/{asset_id}/url": {
      "get": {
        "operationId": "resolve_url_api_v1_admin_assets__asset_id__url_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expires_in_seconds",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 86400,
              "minimum": 1,
              "title": "Expires In Seconds",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedAssetUrlDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Url",
        "tags": [
          "admin",
          "admin-assets"
        ]
      }
    },
...
    "/api/v1/admin/capabilities": {
      "get": {
        "operationId": "list_capabilities_api_v1_admin_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityDTO"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Capabilities",
        "tags": [
          "admin",
          "admin-access"
        ]
      }
    },
...
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "list_roles_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleDTO"
                  },
                  "title": "Response List Roles Api V1 Admin Roles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Roles",
        "tags": [
          "admin",
          "admin-access"
        ]
      },
      "post": {
        "operationId": "create_role_api_v1_admin_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Role",
        "tags": [
          "admin",
          "admin-access"
        ]
      }
    },
    "/api/v1/admin/roles/{role_id}/assign": {
      "post": {
        "operationId": "assign_role_api_v1_admin_roles__role_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Role",
        "tags": [
          "admin",
          "admin-access"
        ]
      }
    },
    "/api/v1/admin/roles/{role_id}/revoke": {
      "post": {
        "operationId": "revoke_role_api_v1_admin_roles__role_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Role",
        "tags": [
          "admin",
          "admin-access"
        ]
      }
    },
...
  },
...
}